requirements.txt
# LegacyLens — RAG System for Legacy Enterprise Codebases
# Python dependencies
# Pin major versions; patch versions float so security fixes apply automatically.

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.29.0,<0.30.0

# ── UI ─────────────────────────────────────────────────────────────────────────
streamlit>=1.35.0,<2.0.0

# ── Vector database ────────────────────────────────────────────────────────────
chromadb>=0.5.0,<0.6.0

# ── Embeddings ─────────────────────────────────────────────────────────────────
voyageai>=0.2.3,<0.3.0

# ── LLM ───────────────────────────────────────────────────────────────────────
openai>=1.30.0,<2.0.0

# ── BM25 fallback retrieval ────────────────────────────────────────────────────
rank-bm25>=0.2.2,<0.3.0

# ── Token counting ─────────────────────────────────────────────────────────────
tiktoken>=0.7.0,<0.8.0

# ── Configuration & env ────────────────────────────────────────────────────────
python-dotenv>=1.0.1,<2.0.0

# ── YAML (eval benchmark) ──────────────────────────────────────────────────────
pyyaml>=6.0.1,<7.0.0

# ── HTTP client (internal API calls) ──────────────────────────────────────────
httpx>=0.27.0,<0.28.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
